# collection-ui/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/publishers/{id}": {
            "get": {
                "tags": [
                    "publisher-controller"
                ],
                "operationId": "getPublisher",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "publisher-controller"
                ],
                "operationId": "updatePublisher",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublisherRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "publisher-controller"
                ],
                "operationId": "deletePublisher",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/designers/{id}": {
            "get": {
                "tags": [
                    "designer-controller"
                ],
                "operationId": "getDesigner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Designer"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "designer-controller"
                ],
                "operationId": "updateDesigner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DesignerRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Designer"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "designer-controller"
                ],
                "operationId": "deleteDesigner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/board-games/{id}": {
            "get": {
                "tags": [
                    "board-game-controller"
                ],
                "operationId": "getBoardGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BoardGame"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "board-game-controller"
                ],
                "operationId": "updateBoardGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardGameRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BoardGame"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "board-game-controller"
                ],
                "operationId": "deleteBoardGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/publishers": {
            "get": {
                "tags": [
                    "publisher-controller"
                ],
                "operationId": "listPublishers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Publisher"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "publisher-controller"
                ],
                "operationId": "createPublisher",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublisherRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publisher"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/designers": {
            "get": {
                "tags": [
                    "designer-controller"
                ],
                "operationId": "listDesigners",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Designer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "designer-controller"
                ],
                "operationId": "createDesigner",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DesignerRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Designer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/board-games": {
            "get": {
                "tags": [
                    "board-game-controller"
                ],
                "operationId": "listBoardGames",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BoardGame"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "board-game-controller"
                ],
                "operationId": "createBoardGame",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardGameRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BoardGame"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/{id}/board-games/{gameId}": {
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "removeFromCollection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "gameId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PublisherRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "name"
                ]
            },
            "Publisher": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "DesignerRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "name"
                ]
            },
            "Designer": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "BoardGameRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "yearPublished": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "minPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "playTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "weight": {
                        "type": "number"
                    },
                    "designerIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "uniqueItems": true
                    },
                    "publisherIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "uniqueItems": true
                    }
                },
                "required": [
                    "title"
                ]
            },
            "BoardGame": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "yearPublished": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "minPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "playTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "weight": {
                        "type": "number"
                    },
                    "designers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoardGameDesigner"
                        },
                        "uniqueItems": true
                    },
                    "publishers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoardGamePublisher"
                        },
                        "uniqueItems": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "BoardGameDesigner": {
                "type": "object",
                "properties": {
                    "designerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "BoardGamePublisher": {
                "type": "object",
                "properties": {
                    "publisherId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
...
        }
    }
}